[FIXED] .sol lint errors/warnings

+++ contracts/oracle/interfaces/IUniswapV2Router.sol
@@ -2,7 +2,6 @@
 
 pragma solidity 0.6.8;
 
-
 /**
  * @title IUniswapV2Router
  * Interface for the UniswapV2 router contract.
@@ -19,6 +18,7 @@ interface IUniswapV2Router {
      * Returns the canonical WETH address.
      * @return The canonical WETH address.
      */
+    // solhint-disable-next-line func-name-mixedcase
     function WETH() external pure returns (address);
 
     /**
@@ -28,12 +28,12 @@ interface IUniswapV2Router {
      * @param amountOut the fixed amount of output asset.
      * @param path An array of token addresses. path.length must be >= 2. Pools for each consecutive pair of addresses must exist and have liquidity.
      */
-    function getAmountsIn(uint amountOut, address[] calldata path) external view returns (uint[] memory amounts);
+    function getAmountsIn(uint256 amountOut, address[] calldata path) external view returns (uint256[] memory amounts);
 
     /**
-     * Receive an exact amount of tokens for as little ETH as possible, along the route determined by the path. The first element of path must be WETH,
-     *  the last is the output token and any intermediate elements represent intermediate pairs to trade through (if, for example, a direct pair does
-     *  not exist).
+     * Receive an exact amount of tokens for as little ETH as possible, along the route determined by the path. The first element of path must be
+     *  WETH, the last is the output token and any intermediate elements represent intermediate pairs to trade through (if, for example, a direct pair
+     *  does not exist).
      * @dev Leftover ETH, if any, is returned to msg.sender.
      * @dev msg.value (amountInMax) is the maximum amount of ETH that can be required before the transaction reverts.
      * @param amountOut The amount of output tokens to receive.
@@ -42,12 +42,17 @@ interface IUniswapV2Router {
      * @param deadline Unix timestamp after which the transaction will revert.
      * @return amounts The input token amount and all subsequent output token amounts.
      */
-    function swapETHForExactTokens(uint amountOut, address[] calldata path, address to, uint deadline) external payable returns (uint[] memory amounts);
+    function swapETHForExactTokens(
+        uint256 amountOut,
+        address[] calldata path,
+        address to,
+        uint256 deadline
+    ) external payable returns (uint256[] memory amounts);
 
     /**
-     * Receive an exact amount of ETH for as few input tokens as possible, along the route determined by the path. The first element of path is the input
-     *  token, the last must be WETH, and any intermediate elements represent intermediate pairs to trade through (if, for example, a direct pair does
-     *  not exist).
+     * Receive an exact amount of ETH for as few input tokens as possible, along the route determined by the path. The first element of path is the
+     *  input token, the last must be WETH, and any intermediate elements represent intermediate pairs to trade through (if, for example, a direct
+     *  pair does not exist).
      * @dev msg.sender should have already given the router an allowance of at least amountInMax on the input token.
      * @dev If the to address is a smart contract, it must have the ability to receive ETH.
      * @param amountOut The amount of ETH to receive.
@@ -57,7 +62,13 @@ interface IUniswapV2Router {
      * @param deadline Unix timestamp after which the transaction will revert.
      * @return amounts The input token amount and all subsequent output token amounts.
      */
-    function swapTokensForExactETH(uint amountOut, uint amountInMax, address[] calldata path, address to, uint deadline) external returns (uint[] memory amounts);
+    function swapTokensForExactETH(
+        uint256 amountOut,
+        uint256 amountInMax,
+        address[] calldata path,
+        address to,
+        uint256 deadline
+    ) external returns (uint256[] memory amounts);
 
     /**
      * Receive an exact amount of output tokens for as few input tokens as possible, along the route determined by the path.
@@ -71,5 +82,11 @@ interface IUniswapV2Router {
      * @param deadline Unix timestamp after which the transaction will revert.
      * @return amounts The input token amount and all subsequent output token amounts.
      */
-    function swapTokensForExactTokens(uint amountOut, uint amountInMax, address[] calldata path, address to, uint deadline) external returns (uint[] memory amounts);
+    function swapTokensForExactTokens(
+        uint256 amountOut,
+        uint256 amountInMax,
+        address[] calldata path,
+        address to,
+        uint256 deadline
+    ) external returns (uint256[] memory amounts);
 }
